# src/swagger/swagger_output.yaml
...
openapi: 3.0.0
info:
  version: 0.0.0
  title: Vinymatic API
  description: ""
host: localhost:5000
basePath: ""
...
schemes: []
securityDefinitions: {}
consumes: []
produces: []
paths:
...
  "/masters":
    get:
      tags:
      - Master
      description: ""
      parameters:
      - $ref: "#/components/parameters/pageParam"
      - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Masters"
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
    post:
      tags:
      - Master
      description: ""
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/requestBodies/AddMaster"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Master"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/masters/search":
    get:
      tags:
      - Master
      - Search
      description: ""
      parameters: 
      - $ref: "#/components/parameters/searchField"
      - $ref: "#/components/parameters/searchValue"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Masters"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/master/{id}":
    get:
      tags:
      - Master
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Master"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
    patch:
      tags:
      - Master
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/requestBodies/AddMaster"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Master"
        "304":
          description: Not Modified
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
  "/master/{id}/versions":
    get:
      tags:
      - Master
      description: ""
      parameters:
      - $ref: "#/components/parameters/idParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VinylsByMaster"
        "401":
          description: Unauthorized
        "404":
          description: Not Found
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
# Transporters ✔︎
...
  "/vinyls":
    get:
      tags:
      - Vinyl
      description: ""
      parameters:
      - name: releaseDate
        in: query
        schema:
          type: string
      - name: country
        in: query
        schema:
          type: string
      - name: label
        in: query
        schema:
          type: string
      - $ref: "#/components/parameters/pageParam"
      - $ref: "#/components/parameters/limitParam"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vinyls"
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
    post:
      tags:
      - Vinyl
      description: ""
      requestBody:
        content:
          multipart/form-data:
            schema:
              $ref: "#/components/requestBodies/AddVinyl"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Vinyl"
        "400":
          description: Bad Request
        "401":
          description: Unauthorized
        "500":
          description: Internal Server Error
      security:
      - bearerAuth: []
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  parameters:
...
    pageParam:
      name: page
      in: query
      schema:
        type: integer
    limitParam:
      name: limit
      in: query
      schema:
        type: integer
    searchField:  
      name: withCol[field]
      description: Permet de rechercher sur une colomne en particulier avec le __nom de colomne__( ___field___ ) souhaité
      example: name
      in: query
      required: true
      schema:
        type: array
        items:
          type: string
    searchValue:  
      name: withCol[value]
      description: Permet de rechercher sur une colomne en particulier avec la __valeur__( ___value___ ) souhaitée
      example: michael
      in: query
      required: true
      schema:
        type: array
        items:
          type: string
...
    idParam:
      name: id
      in: path
      required: true
      schema:
        type: integer
...
  requestBodies:
...
    AddMaster:
      type: object
      properties:
        title:
          type: string
        description:
          type: text
        releaseDate:
          type: string
        images:
          type: array
          items:
            type: string
            format: binary
        thumbnail:
          type: string
          format: binary
        # tracklist:
        #   type: object
      # required:
      # - title
      # - releaseDate
...
    AddVinyl:
      type: object
      properties:
        title:
          type: string
        description:
          type: text
        country:
          type: string
        releaseDate:
          type: string
        images:
          type: array
          items:
            type: string
            format: binary
        thumbnail:
          type: string
          format: binary
        identifiers:
          type: array
...
  schemas:
...
    MinOfArtist:
      type: object
      properties:
        id: 
          type: integer
        idArtist: 
          type: integer
        name:
          type: string
        thumbnail:
          type: string
        artistUrl:
          type: string
        ressourceUrl:
          type: string
...
    MinOfGenre:
      type: object
      properties:
        id: 
          type: integer
        name: 
          type: string
...
    MinOfMaster:
      type: object
      properties:
        id: 
          type: integer
        idMaster:
          type: integer  
        title:
          type: string
        releaseDate:
          type: string    
        thumbnail:
          type: string
        masterUrl:
          type: string
        resourceUrl:
          type: string
...
    MinOfStyle:
      type: object
      properties:
        id: 
          type: integer
        name: 
          type: string
...
    MinOfVinyl:
      type: object
      properties:
        id: 
          type: integer
        idRelease: 
          type: integer  
        title:
          type: string  
        thumbnail:
          type: string  
        vinylUrl:
          type: string
        resourceUrl:
          type: string
  
  # Array of
    ArrayOfArtist:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfArtist"
    ArrayOfAudio:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfAudio"
    ArrayOfFormat:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfFormat"
    ArrayOfGenre:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfGenre"
    ArrayOfLabel:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfLabel"
    ArrayOfMaster:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfMaster"
...
    ArrayOfStyle:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfStyle"
    ArrayOfUser:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfUser"
    ArrayOfVinyl:
      type: array
      items: 
        $ref: "#/components/schemas/MinOfVinyl"
...
    Master:
      type: object
      properties: 
        master:
          allOf:
          - $ref: "#/components/schemas/MinOfMaster"
          - type: object
          properties:
            description:
              type: text
            releaseDate:
              type: string
            masterUrl:
              type: string
            resourceUrl:
              type: string
            images:
              type: array
              items:
                type: string
            tracklist:
              type: array
              items:
                type: object
                properties:
                  name: 
                    type: string
                  duration:
                    type: integer
            Vinyls:
              $ref: "#/components/schemas/ArrayOfVinyl"
            MasterMainArtists:
              $ref: "#/components/schemas/ArrayOfArtist"
            MasterCredits:
              $ref: "#/components/schemas/ArrayOfArtist"
            Genre:
              $ref: "#/components/schemas/ArrayOfGenre"
            Style:
              $ref: "#/components/schemas/ArrayOfStyle"
...
    Vinyl:
      type: object
      properties: 
        vinyl:
          allOf:
          - $ref: "#/components/schemas/MinOfVinyl"
          - type: object
          properties:
            description:
              type: text
            country:
              type: string
            releaseDate:
              type: string
            vinylUrl:
              type: string
            resourceUrl:
              type: string
            images:
              type: array
              items:
                type: string
            tracks:
              type: string
            identifiers:
              type: array
              items:
                type: object
                properties:
                  type:
                    type: string
                  value:
                    type: string
                  description:
                    type: string
            Masters:
              $ref: "#/components/schemas/MinOfMaster"
            Audio:
              $ref: "#/components/schemas/ArrayOfAudio"
            VinylMainArtists:
              $ref: "#/components/schemas/ArrayOfArtist"
            VinylCredits:
              $ref: "#/components/schemas/ArrayOfArtist"
            VinylLabels:
              $ref: "#/components/schemas/ArrayOfLabel"
            VinylSocietes:
              $ref: "#/components/schemas/ArrayOfLabel"
            Format:
              $ref: "#/components/schemas/ArrayOfFormat"
            Genre:
              $ref: "#/components/schemas/ArrayOfGenre"
            Style:
              $ref: "#/components/schemas/ArrayOfStyle"
            Sells:
              $ref: "#/components/schemas/ArrayOfUser"
            VinylLike:
              $ref: "#/components/schemas/ArrayOfUser"
...
    Masters:
      type: object
      properties:
        masters:
          allOf:
          - $ref: "#/components/schemas/ArrayOfMaster"
...
    Vinyls:
      type: object
      properties:
        vinyls:
          allOf:
          - $ref: "#/components/schemas/ArrayOfVinyl"
    VinylsByMaster:
      type: object
      properties:
        master:
          type: object
          properties:
            id: 
              type: integer
            title:
              type: string
            Vinyls:
              allOf:
              - $ref: "#/components/schemas/ArrayOfVinyl"
...